tests/contract/test_slack_bot_mentions.py
"""
Contract test for Slack Bot mention events.

This test validates that our Slack event handler correctly processes
bot mention events according to the Slack Events API specification.

Reference: specs/002-slack-bot-ai/contracts/slack_events.yaml
"""

from typing import Any, Dict

import pytest
from fastapi.testclient import TestClient

# NOTE: These imports will fail until implementation exists (TDD)
# This is expected and required for TDD approach
try:
    from src.integrations.slack_handler import SlackEventHandler
    from src.main import app
except ImportError:
    # Expected during TDD phase - tests should fail first
    pytest.skip("Implementation not yet available", allow_module_level=True)


class TestSlackBotMentions:
    """Test Slack bot mention events contract."""

    @pytest.fixture
    def client(self) -> TestClient:
        """Create test client for Slack event endpoint."""
        return TestClient(app)

    @pytest.fixture
    def bot_mention_payload(self) -> Dict[str, Any]:
        """Valid bot mention event payload."""
        return {
            "type": "event_callback",
            "team_id": "T1234567890",
            "api_app_id": "A1234567890",
            "event": {
                "type": "app_mention",
                "text": "<@U0123456789> 来週チームでランチしませんか？ <@U1111111111> <@U2222222222>",
                "user": "U0123456789",  # Organizer
                "channel": "C1234567890",
                "ts": "1234567890.123456",
                "event_ts": "1234567890.123456"
            }
        }

    @pytest.fixture
    def bot_mention_with_thread_payload(self) -> Dict[str, Any]:
        """Bot mention event in thread."""
        return {
            "type": "event_callback",
            "team_id": "T1234567890",
            "api_app_id": "A1234567890",
            "event": {
                "type": "app_mention",
                "text": "<@U0123456789> スケジュールの候補日程をお願いします",
                "user": "U0123456789",
                "channel": "C1234567890",
                "ts": "1234567890.123456",
                "thread_ts": "1234567890.123000",  # Parent message timestamp
                "event_ts": "1234567890.123456"
            }
        }

    @pytest.fixture
    def bot_mention_with_participants_payload(self) -> Dict[str, Any]:
        """Bot mention with @here/@channel participants."""
        return {
            "type": "event_callback",
            "team_id": "T1234567890",
            "api_app_id": "A1234567890",
            "event": {
                "type": "app_mention",
                "text": "<@U0123456789> 勉強会を企画します <!here>",
                "user": "U0123456789",
                "channel": "C1234567890",
                "ts": "1234567890.123456",
                "event_ts": "1234567890.123456"
            }
        }

    def test_bot_mention_success(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test successful bot mention event processing."""
        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": "1234567890"
            }
        )

        # Should acknowledge event quickly
        assert response.status_code == 200
        assert response.text == "OK"

    def test_bot_mention_thread_event(
        self,
        client: TestClient,
        bot_mention_with_thread_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention in thread (intermediate confirmation)."""
        response = client.post(
            "/slack/events",
            json=bot_mention_with_thread_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": "1234567890"
            }
        )

        # Should process thread messages for confirmations
        assert response.status_code == 200
        assert response.text == "OK"

    def test_bot_mention_with_here_mention(
        self,
        client: TestClient,
        bot_mention_with_participants_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention with @here participant specification."""
        response = client.post(
            "/slack/events",
            json=bot_mention_with_participants_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": "1234567890"
            }
        )

        # Should handle @here/@channel mentions
        assert response.status_code == 200
        assert response.text == "OK"

    def test_bot_mention_missing_required_fields(self, client: TestClient) -> None:
        """Test bot mention with missing required fields."""
        invalid_payloads = [
            # Missing user field
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "text": "<@U0123456789> test",
                    "channel": "C1234567890",
                    "ts": "1234567890.123456"
                }
            },
            # Missing channel field
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "text": "<@U0123456789> test",
                    "user": "U0123456789",
                    "ts": "1234567890.123456"
                }
            },
            # Missing text field
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "user": "U0123456789",
                    "channel": "C1234567890",
                    "ts": "1234567890.123456"
                }
            }
        ]

        for payload in invalid_payloads:
            response = client.post(
                "/slack/events",
                json=payload,
                headers={
                    "Content-Type": "application/json",
                    "X-Slack-Signature": "v0=test_signature",
                    "X-Slack-Request-Timestamp": "1234567890"
                }
            )

            # Should handle invalid payloads gracefully
            assert response.status_code in [200, 400]

    def test_bot_mention_invalid_signature(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention with invalid Slack signature."""
        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=invalid_signature",
                "X-Slack-Request-Timestamp": "1234567890"
            }
        )

        # Should reject invalid signatures
        assert response.status_code == 401

    def test_bot_mention_missing_signature(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention without Slack signature headers."""
        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={"Content-Type": "application/json"}
        )

        # Should require signature verification
        assert response.status_code == 401

    def test_bot_mention_old_timestamp(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention with old timestamp (replay attack protection)."""
        import time
        old_timestamp = str(int(time.time()) - 400)  # 400 seconds ago

        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": old_timestamp
            }
        )

        # Should reject old timestamps (> 5 minutes)
        assert response.status_code == 401

    @pytest.mark.asyncio
    async def test_bot_mention_response_time(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test bot mention response time compliance."""
        import time

        start_time = time.time()

        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": str(int(time.time()))
            }
        )

        end_time = time.time()
        response_time = end_time - start_time

        # Slack requires acknowledgment within 3 seconds
        assert response_time < 3.0
        assert response.status_code == 200

    def test_bot_mention_contract_compliance(
        self,
        client: TestClient,
        bot_mention_payload: Dict[str, Any]
    ) -> None:
        """Test contract compliance with OpenAPI specification."""
        response = client.post(
            "/slack/events",
            json=bot_mention_payload,
            headers={
                "Content-Type": "application/json",
                "X-Slack-Signature": "v0=test_signature",
                "X-Slack-Request-Timestamp": str(int(__import__("time").time()))
            }
        )

        # Validate response matches contract spec
        assert response.status_code == 200
        assert response.text == "OK"

        # Response should be plain text
        assert response.headers.get("content-type", "").startswith("text/plain")

    def test_bot_mention_japanese_text_handling(self, client: TestClient) -> None:
        """Test bot mention with Japanese text and event types."""
        japanese_payloads = [
            # Dining event
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "text": "<@U0123456789> 今度チームでランチしませんか <@U1111111111>",
                    "user": "U0123456789",
                    "channel": "C1234567890",
                    "ts": "1234567890.123456"
                }
            },
            # Study session
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "text": "<@U0123456789> 来週勉強会をしたいです <!channel>",
                    "user": "U0123456789",
                    "channel": "C1234567890",
                    "ts": "1234567890.123456"
                }
            },
            # Meeting
            {
                "type": "event_callback",
                "event": {
                    "type": "app_mention",
                    "text": "<@U0123456789> MTGの予定を調整してください",
                    "user": "U0123456789",
                    "channel": "C1234567890",
                    "ts": "1234567890.123456"
                }
            }
        ]

        for payload in japanese_payloads:
            response = client.post(
                "/slack/events",
                json=payload,
                headers={
                    "Content-Type": "application/json",
                    "X-Slack-Signature": "v0=test_signature",
                    "X-Slack-Request-Timestamp": str(int(__import__("time").time()))
                }
            )

            # Should handle Japanese text correctly
            assert response.status_code == 200
            assert response.text == "OK"